method overriding in java

+++ core_java_concepts_projects/src/com/amdocs/MethodOverridingTest.java
@@ -15,7 +15,7 @@ public class MethodOverridingTest extends Parent {
 	}
 	
 	@Override
-	protected long getNumbericValue() {
+	protected short getNumbericValue() {
 		System.out.println("Message from MethodOverridingTest class");
 		return 2L;
 	}
